LUIS model definition and tests for v0.1 Simple and v0.2 Extended
LUIS model exports for v0.1 Simple Item ordering and v0.2 Extended Classification ordering (names with numerics and ordering using aliases.
For each version there is a json file with LUIS Batch Tests and an outline file for the E2E pipeline testing which will test for entity extraction and mapping to the correct DSEs.